Fixed truncation warnings in the ContentBrowser modules
#jira UE-166269
#rb yohann.dossantos
#preflight 63e12ffa797b029c0a6507fe

[CL 24032337 by JeanMichel Dignard in ue5-main branch]

+++ Engine/Source/Editor/AssetDefinition/Public/AssetDefinition.h
@@ -57,7 +57,7 @@ struct FAssetArgs
 	template<typename ExpectedObjectType>
 	TArray<ExpectedObjectType*> LoadObjects(const TSet<FName>& LoadTags = {}, TArray<FAssetData>* OutAssetsThatFailedToLoad = nullptr) const
 	{
-		FScopedSlowTask SlowTask(Assets.Num());
+		FScopedSlowTask SlowTask((float)Assets.Num());
 	
 		TArray<ExpectedObjectType*> LoadedObjects;
 		LoadedObjects.Reserve(Assets.Num());
