+++ Engine/Source/Editor/ContentBrowserData/Public/ContentBrowserVirtualPathTree.h
@@ -48,9 +48,9 @@ public:
 	/** Recursively enumerates all child paths from the specified base path relative to this node */
 	bool EnumerateSubPaths(FName BasePath, TFunctionRef<bool(FName, FName)> Callback, bool bRecurse = true) const;
 
-	uint32 GetAllocatedSize(void) const
+	SIZE_T GetAllocatedSize(void) const
 	{
-		uint32 AllocatedSize = ParentPathToChildPaths.GetAllocatedSize() + ChildPathToParentPath.GetAllocatedSize() + VirtualToInternalMounts.GetAllocatedSize();
+		SIZE_T AllocatedSize = ParentPathToChildPaths.GetAllocatedSize() + ChildPathToParentPath.GetAllocatedSize() + VirtualToInternalMounts.GetAllocatedSize();
 
 		for (const TPair<FName, TSet<FName>>& Pair : ParentPathToChildPaths)
 		{
